Move Xorable from routing_table to xor_name
