+++ src/lib.rs
@@ -187,10 +187,10 @@ pub use crate::{
     node::{Node, NodeBuilder},
     pause::PausedState,
     routing_table::RoutingTableError,
-    routing_table::{Authority, Prefix, Xorable},
+    routing_table::{Authority, Prefix},
     types::MessageId,
     utils::XorTargetInterval,
-    xor_name::{XorName, XorNameFromHexError, XOR_NAME_BITS, XOR_NAME_LEN},
+    xor_name::{XorName, XorNameFromHexError, Xorable, XOR_NAME_BITS, XOR_NAME_LEN},
 };
 #[cfg(feature = "mock_base")]
 pub(crate) use chain::Chain;
