+++ src/routing_table/authority.rs
@@ -6,8 +6,8 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::{Prefix, Xorable};
-use crate::xor_name::XorName;
+use super::Prefix;
+use crate::xor_name::{XorName, Xorable};
 use std::fmt::{self, Binary, Debug, Display, Formatter};
 
 /// An entity that can act as a source or destination of a message.
