+++ src/routing_table/prefix.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::xorable::Xorable;
+use crate::xor_name::Xorable;
 use std::{
     cmp::{self, Ordering},
     fmt::Result as FmtResult,
