+++ src/xor_name/mod.rs
@@ -6,7 +6,9 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::routing_table::Xorable;
+mod xorable;
+
+pub use self::xorable::Xorable;
 use hex::{FromHex, FromHexError, ToHex};
 use num_bigint::BigUint;
 use rand;
@@ -261,7 +263,6 @@ impl AsRef<XorName> for XorName {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::routing_table::Xorable;
     use maidsafe_utilities::serialisation::{deserialise, serialise};
     use rand;
     use std::cmp::Ordering;
